phase9-04: 평균 생산시간 타입을 double로 변경, 0 이상 허용
- Sample: avgProductionTime int → double, 검증 조건 1이상 → 0이상
- ProductionJob: totalProductionTime int → double (avgProductionTime 연동)
- SampleService.register: avgProductionTime 파라미터 int → double
- ConsoleIO: readDouble() 메서드 추가
- SampleMenuHandler: readDouble 사용, 출력 포맷 %12d → %12.1f
- ProductionMenuHandler: 출력 포맷 %14d → %14.1f
- JsonDataStore.SampleDto: avgProductionTime int → double
- SampleTest: 0 허용 / 소수 허용 / 음수 예외 테스트로 업데이트 (90 → 92개)

+++ src/main/java/org/example/domain/ProductionJob.java
@@ -4,7 +4,7 @@ public class ProductionJob {
 
     private final Order order;
     private final int actualProductionCount;
-    private final int totalProductionTime;
+    private final double totalProductionTime;
 
     public ProductionJob(Order order, int shortfall) {
         if (shortfall < 1) {
@@ -17,5 +17,5 @@ public class ProductionJob {
 
     public Order getOrder() { return order; }
     public int getActualProductionCount() { return actualProductionCount; }
-    public int getTotalProductionTime() { return totalProductionTime; }
+    public double getTotalProductionTime() { return totalProductionTime; }
 }
